+++ Documents/.$Architecture.drawio.bkp
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2023-03-10T15:56:13.308Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.6.1 Chrome/106.0.5249.199 Electron/21.3.3 Safari/537.36" version="20.6.1" etag="7jV0IcmoKbJRnyOljCgs" type="device"><diagram id="UQbhMwpoYKBb8NZh9liM" name="Page-1">7V1Zd6M2FP41fvQcs9uPiZdJTjOZdJJ22qceGWSsFiMq5CTur69YbZBsAzYGJzzMmXBBC9+9upsuck8Zr96/EuAtv2ELOj15YL33lElPluXhYMD+CyibiKLqmhpRbIKsiDbYEp7RfzAiSgl1jSzox7SIRDF2KPKyRBO7LjRphgYIwW/ZxxbYsTIED9iQIzybwOGpP5FFl/GLKcpoe+MOInsZD60oySuvQPJ0TPCXwMJvOyRl2lPGBGMa/bV6H0MngC8LzGzP3XRmBLq0SAP0MPV/Ue8QVsbP88XzzfT7j2U/5oZPN8kbQ4sBEF9iQpfYxi5wplvqLcFr14JBrwN2tX3mAWOPESVG/BtSuom5CdYUM9KSrpz4Lpsw2fwRtw8v/ty9mLxnrgLR0eQvWkJ4ggStIIUkfip6g2Dae4GJST5eExMeQCORMEBsSA88J6fsY5IPMZsN2bB2BDqAotfsPEAsgXb63JZH7I+YTSVYpkkt4tmWLQKuBVcL7NK4Uzm49hm49CZYmYxiOsD3kZmQZ8hJhmuAqeqwSa7G/b4CZx2P9BPObzzPQSYbHbs9WXfY7G/nJMN6/d91oD9CmPt+iPMNe4Ah9R6imNxnf9nB/8D3XEhNTCAnRFkReVsiCp89EEL7xrR7VhzybI1nDwmF74c5xgOcWItRrChjWyEPlC+GFpHetrpXMuLHljtqdzioiS0yx5YJe9dBn/37tnn+9YFDkWl5L/jT3DiIwUmU41jOI+Af5ikBmP/YITu+rynrBsb0iL8TSTsP4KqeBVzRhIDrl8R7yOHdhLI7o/bRC2ofo0nlo3OoP2AbuaepiHNIqJyVUCORu119IAvkU69LPhO/9OMIqFFQQEeNWkdeD38S2KU9K+RCuKufFnelSdwNDvY71nPj+lhvnz7WrlweT1WrcdMnjNjIKaMMKcsoXcoxIJL/uFWOB+k0TmDLh/PjRkX1RqOOnPxpcZcbxV35cG5hYdwb9QtHHOxjYAGfEuw3biyHeR1sNG0sE6H8OFIqFc7YNupGyx/OjS4OfKN+tMxnNj4N8FqTwCfT3AH+cdY6lTwcNa2SozAzm2AG/nKOAbGYCdPBKkDBnfteikCD+I2M1sV//AIfO4i9DmzeAdC1LFqa1jhafF7hGZJX1BsrvRtD4DMdASmLaA0LtHnIZF6RtcfF1Funz2Ren90hnwYCdqsQZOLGMcuLmDRQGgdN40B7YKABG64ah0saDFonY7wWG+MVcK0WrEh12LqgTxQjU2hjgkDzeEmS1jbAlALBWrlaCeQ4Y+zgoGDJxcFu/lnKJXK+WLqbvwvcQBUgZ+STweeDjldk+SqWHJDs/WkWrcCy/gNzeAkgBA6yXXZpBs4eo98GaLJhnJv4xgpZVhi+idiTZWBYzRJPqqbqFiXv2Yisjqi0xaiLWSqfK/45DYpbQpeQgcVej8AFRHRNwHlDkRrw1bjqIQG+mgDfofFFqwlhTeYQfgKUSStbB4PJZNLm1bDDnd3VoZyBVXJO5UsCJ18WsUqpaylovN5K6u2WUlI69zuCb35CZsNs7yREb4eQWy2zp/WcKcG7gLPbLrxDPYQE3wNuRkaSgr6QjX0/4uNNtFZ7ooq/XpADU0Y7fbKBo26TB2Zxtc+2wjA7ifbM9GJD3bP18P7F9Jc7DGszMLOoPiAvdu2feMfRfRwN8qWtm+YFZeLK+XexoYJUweOsQ4YfKsg8YbJpMzjNrafpCl2X2Mx+QJsxFJL2Te2SKETbG/41cOwjLhuRGfoMiAsEEa+CpOsnl8TOoW0dMOwyTmF1stnJZruAubCKjjd6uoXwCYZqt2cyS8uaOlG8YkU5m2EyR5YFu3RxTWzs8PzwQx1j8fFdQTYZ5NrsSttevYTFyH153/4rfoVk4YTnbiyjFXx0429nn5G92WKxkM3gsA6bACtIQmTuWSNd7+U3LNkdS5/rmp6+12nVIspQzmwd9qXBkNs7TGtwdvcOJdlQa9o91PkSwYSv/B7hXjHbHvrQNyP0AllDLqIIOEJ5GzN2Eew4kPhHBE64W2mh14QU7v9tu2OKbKflznOCpsG+U7WWqUNQrfnjrOKEo6RwtcZJzqNa6xe2kE/roQYpSg36vhntEajMrDqNtU9jjXJ1P0xjjYpqLF2vS2PxhSnnLeipp+RNUnKVI/30nLhMlU9iJDJopmeJnB9NvmDweqreIqYpci9fBXcOdo3UnOwnnNmt9FEEsl9bpY9epNKnBi0bHo14FjNdxlz85pc0b+HeUpkWSaK11ChpVqqcva9iqCs5JFUsstCnv7gxPzibzpCf8XQ6SRGYHqEdV+pTZqKvW/PKbPLy/XDVYmaNsafrkf5OsPYIllFdsOrzaPgvbHjBmrrrVWHJetl4sLx1CVo9Ecx4Bg8060Rrn84ajrKipQu+3bq0aBU4oaZq6BF5sfV8baPqWSjTY50zn9skgVXmAwOtNiz5T7smeAUEBze2L+BQe+cPOJTcp4qSXPTz4dpMtCGKtPOa9Af0sI9oZqP2mEK9BT5M223KBQVrQBCu1jjS4dXaJrFCxZHTqKFa+8dZtXadi3OOtH3+ix+1cUNkFPFxJoCCoisyeDbIoTK2d85KlUxpXnuLPpG8tJBcp7fCJUolZSAJ3BURmPW5Kwbvrty7CwIgk6u1SdeEP+D0c/gtkqrk2VX0kIXaHJdhkURpvszymIp8hr6PcKkU5o2HyseNXx08B8703YQeLTneeO1TvBoDc8nUbucnXNQE5P334Kvupi3AsEiO7T7QMQtQYiXcV/ff7zkHviePy7Q/zQ+/P9URvy/oiXfLpOgJRi1I6wyLeNPll4lYfEMjUmXn6T7S7g/YtiGp1MHUNcnGo5XapqasUuucTSvcR1FQYpNXaW4z5NBqgHZW83R1wFlNwUlAl1YH1xk3yblfRGKqVZTlHYnCJrU2LPmwaet3tz5e0no15Hk5NgkMoOhnq2qLlpLxD5q/J0DCjXt5cPwomd2gacmAsUpr1AvUM//1ADZ4TYvUvbdtzt8Xi+ua9jfm9QCbhQz+2jkj4hWaPmLaZpEMT8Pww5CobvZ29Tnt8YK434rrSwO1mBskydqwLrtQ5NOAO+h42Rr+g/YgcrbLONlx0FI+DRYFSl2Mfnrpxfao5jqccna5/R3p8N7X7e9xK9P/AQ==</diagram></mxfile>
