+++ Documents/.$Architecture.drawio.dtmp
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2023-03-10T16:20:39.554Z" agent="5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.6.1 Chrome/106.0.5249.199 Electron/21.3.3 Safari/537.36" etag="05671OvrgV36RZpeIza6" version="20.6.1" type="device"><diagram id="UQbhMwpoYKBb8NZh9liM" name="Page-1">7V1bd6M2EP41fvQeAwbsx8SXTU6z2XSz7bZPPTLIWC1GVMhJ3F9fcTUg2QZibHnDw54Ng258M5oZjUZyT5us3z4T4K++YBu6PXVgv/W0aU9VVc0csv9CyjamDHU1oTgE2TFtsCM8o/9gTFRS6gbZMEhoMYli7FLkF4kW9jxo0QINEIJfi8WW2LULBB84kCM8W8DlqT+QTVfph2nj3Ys7iJxV0rWmDZKRr0FaOiEEK2Dj1xxJm/W0CcGYxn+t3ybQDeErAjPf8zYbGYEerVIBPcyCX4Z3CGuT58Xy+Wb29duqn3AjoNv0i6HNAEgeMaEr7GAPuLMd9ZbgjWfDsNUBe9qVecDYZ0SFEf+GlG4TboINxYy0oms3ecsGTLZ/JPWjhz/zD9O3wlMoOrr6SU8JT5CgNaSQJKV4HBJoArwhFjzw8alAAeJAeqCcGpcLkcl1kKD8GWI2GrJlBQh0AUUvRdEBiQQ6Wbkdj9gfCZtqsExXJOLZji0CroVPS+zRpFE1fA4Y2vQmnJmMYrkgCJCVkufITbtrn6nDkVRcTdp9Ae4m6ekHXNz4voss1jv2eqrhss+5XZAC641/N6H+iGDuBxHON6wAg+4tQjF9z/5ywv9B4HuQWphAToiKIvK6QhQ++yDC+pVp96I4lNm6l18vkFD4dhDg1FqME0WZ2Ap1oH0y9Zj0utO9ipkUW+XU7mjQEltUji1T9q2DPvv3Zfv86wOHItPyfvintXURg5Nox7FcxMA/LDICsP5xInZ83VDWDEzoMX+nin4awIdGEXBNFwJunBPvEYf3JZRdc+1jVNQ+plTKx+BQf8AO8t6nIk4hoWpRQs1U7vL6QBXIp9GWfKZ+6dUKqFlRQMdSCajC6+GfE3ZFkQv34UfBXZMKd5OD/Y61fHF9bMinj/XrkseTq9Wk6hNGbCgZo0ylyChDKTEgnhBJrRIPsmG8gy3X7seNq+oNuRw59aPgrsqFu3btbmFl3OXyC8cc7BNgg4ASHFzcWI7KOti8tLFMhfJqpTT1jo+LqVxutHrtbnR14OXyo1U+svGzAq9LBXw67hzwj3PpVPJofGmVnK478wFmEKwWGBCbmTADrEMUvEXgZwhcEL+xKd36j5/gExexL4SXdwAMvYiWrl8cLT6u8AzJC+pNtN6NKfCZjoBURLSFCXp5yFRekcnjYhrS6TOV12d3KKChgN1qBFn44piVRUwZaBcHTedAe2CgAQeuLw6XMhhIJ2O8FpvgNfBsCWbkcCTdok+0RqbQwQSBy+OlKLpsgGkVFmv1ciWQ606wi8OEJQ+Hu/knSZco+WLZbn4euMFQgJxZDgafDjpekZWzWEpAso+lRbRCy/oPLOElgBC4yPHYoxU6e4x+G0LHunFvkhdrZNvR8k3EniIDo2yWZFAtZbdoZc9GZHVEqS1mW8wa8rHiH7MwuSVyCRlY7IsJXEJENwScdinSAr46lz0kwFcX4DsyP+ktIayrHMJPgDJpZfNgMJ1OZZ4NOe7kZ4d2Cs1VUvmKwMlXRazS2poKOq+30ny7lZKmzv2O4GuQklk3uzcp0c8RSrNl/rRZMCV4F3J214R/qIWIEPjAK8hImtAXsbEfxHy8iedqT5Tx1wtjYNo41ybrOG42LTBPsn12GYbFQcgz0rN1dc/mw9snK1jlGCYzMPM4P6AsdvIPvOPoPo6G8VLphnlGmbhy/p2tqzBU8DjvkOG7CiNPmGxlBudy82m2RtclNvNv0GEMhUS+oZ0ThXh7I7gGjv2M00Zkhj4C4gJBxOsw6PrBJbFzaKUDhj0mIaxONjvZlAuYM6voZKOnmwgfoCu5PZN5ltbUieIVK8r5HJMFsm3YhYtbYmOH50/f1TEWH98VZINBnsOe9N3T9ygZua/u23/FL5As3ejejVU8g49u/OX2GdmXLZdL1Qov63AIsMMgROGdPTaMXnnDkr2xjYWhG9l3vS9bRBupha3DvjIYcXuHWQ5Ofu9QUdO7iE5/ep5PEUz5yu8R7hWz3aUPfStGL5Q15CGKgCuUtwljF8GuC0lwROCEu5U2eklJ0f7frjmmyHI1c+UEVcN9p2Y1M4egWfXHecMBx0HhZpXTmEez2t/ZRH5fCy1IUWbQ941oj0AVRtVprH0aa1zK+2Eaa1xVYxlGWxqLT0w5bUJPOylvilbKHOln98QVsnxSI1FAM7tL5PRo8gmD15P1FjNNU3vlLLhTsGs8LMl+ypl8po8mkP3WMn2MKpk+LWjZ6GrEk5jpOubit6CmeYv2lurUSAKttTqJg1Lbesa+gZVu5Iw0scZCf/7shvzgaDojfsKb6RRNYHaENlxrT5GJTraWFdn0+9fDGYuFKcZKtyP9nWDtESyzuWC1583wp2t4wZp5m3Vlyfq+9WF9yxLWeiKY8eyQdelEa5/OGo2LomUIzm2dW7Qq3E7TdNkRe7DtnLTRim6skl3pXDhqk8aXCocL9Naw5I91TfEaCC5tlG+xMeydfrGhlY4pKmrVo8OtmWhTtMoua9Jv0McBooVN2mMK9RYEMKtXy4/+LdgAgnCzyrEOb1Y3Tcho2HO2j92s/uO8Wb3OxTlFyL582md4cUNkVvFxpoCCqjMyLBvGTxnbO2elSZS0rL1FxyPPLSTX6a1wQVJFGygCd0UEZnvuism7K/fekgDI5Gpj0Q3hLzf9GH6LMtTK7Kp6wUJrjsuoSpC0nGJ5TEU+wyBAuFb48sZH9deNn128AO7szYI+rdnfZBNQvJ4Aa8XUbucnnNUElP338ET3pS3AqEqM7T7UMUtQYybcN/ff7zkHvqdO6tR/nx9+/15H/L6iJ95Nk6q3F0kQ1hlV8abrTxOx+EZGJJXierIbafcH7DiQNGpg5llk69NGdTNT1qh2yaZVbqMqKInJazS2OXJpM0A7q/l+dcBZTcEtQOdWB9e5blJLv4bEVKsoyjsWLZuGrWHJL5t2frf06yW910Kcl2OTwACKfrKqtdVS2v9B8/cESLRxrw6OXyOTXzStGFZ2bY16hlzmvx7AFm9olZx32cb8dbm8rmF/YV4PcNiSIdi4J0S8QdVHTGUWyegmjCBaErXN3i4/Rx4viPuduL4yGFZzgxRVH7VlF6ocC7iDrl/M3z9oD2Jnu46TnSxa6ofB4oVSt0Z/f+rF7prmNpxy9rj7Deno3efdb3Frs/8B</diagram></mxfile>
